Add missing world query params (#472)
* feat(world_search_params): Add missing params

"avatarSpecific" and "notag" are used in some places for searching for
worlds. `notag: exlcude worlds with these tags` `avatarSpecific: only
include avatar worlds`

Another field is "organization". Ive only ever seen this set as "vrchat"
though. I didnt include it because of that, but thought id mention it
just in case

* feat(active_world_params): add avatarSpecific

I missed adding the avatarSpecific parameter to the worlds/active endpoint